Add tool to sync session videos with YouTube playlists
This tool can use templating to map session names/times/whatevers to
filenames matching session recordings, and then upload those files to
youtube and build a playlist for them. Session descriptions and speakers
are also copied to the playlist, with the formatting of the title etc
controlled by jinja2 templates.

The source for the session data is the "json schedule dump" saved or
downloaded using a token URL.