Make it possible to re-register with the same email address
This moves the constraint that an email has to be unique per conference
to only applying to non-canceled registrations, making it possible to
cancel a registration and then restart the process with a new one if
necessary, while logged in as the same user. This can for example be
used to cancel a general registration for someone who then wants to
register as a speaker, and can use the same email and userid.

The canceled registration will remain, but have the attendee set to NULL
and move the attendee to registrator, meaning we can still track it and
the attendee can still see it themselves, making it possible to access
things like invoices after the completion.

The cross conference email is set up to handle this case without sending
duplicates email. There are likely some other places around that may
need some adjustments that haven't been found yet.