Add support for cancel requests
This makes it possible for an attendee to request a cancelation of their
registration using a button on the registration dashboard (while signed
in, which means it's known who the attendee is, and there is no need to
verify emails). When they do so, a cancelation request is recorded and
visible on the admin dashboard and an notification is sent to the
admins. The actual cancelation still has to be done at this point, which
gives the admin a chance to double check that refund patterns were OK
and override if necessary.